+++ cmd/root.go
@@ -39,7 +39,7 @@ func init() {
 	rootCmd.PersistentFlags().StringVarP(&flagFile, "file", "f", "stack.yaml", "YAML stack file")
 	rootCmd.PersistentFlags().StringVar(&flagRuntime, "runtime", "", "container runtime: docker|podman (overrides YAML)")
 	rootCmd.PersistentFlags().StringVar(&flagSocket, "socket", "", "override runtime socket path")
-	rootCmd.PersistentFlags().StringVarP(&flagOutput, "output", "o", "text", "output format: text|json|yaml")
+	rootCmd.PersistentFlags().StringVarP(&flagOutput, "output", "o", "console", "output format: console|json|yaml")
 	rootCmd.PersistentFlags().BoolVar(&flagNoColor, "no-color", false, "disable ANSI color output (also respects NO_COLOR env var)")
 	rootCmd.PersistentFlags().StringVar(&flagProject, "project", "", "override project name from YAML")
 }